# apps/api/config/swagger.yaml
openapi: 3.0.2
info:
  title: Ukraine Air Raid Alert API
  description: ""
  contact:
    email: pm.antondolia@gmail.com
  version: "1.0"
tags:
  - name: Alerts
  - name: Regions
  - name: Render
paths:
  /api/regions:
    get:
      tags:
        - Regions
      summary: Return all registered regions 
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Region'

  /api/status:
    get:
      tags:
        - Alerts
      summary: Return alert statuses of all regions
      parameters:
        - name: short
          in: query
          required: false
          description: If true will be use ShortStatus schema
          schema:
            type: boolean

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Status'

  /api/status/{region_id}:
    get:
      tags:
        - Alerts
      summary: Return region status
      parameters:
        - name: region_id
          in: path
          required: true
          description: Id of region
          schema:
            type : integer
            format: int64
            minimum: 0
        - name: short
          in: query
          required: false
          description: If true will be use ShortStatus schema
          schema:
            type: boolean
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
        '404':
          description: Error region with given id not found
          content:
            application/json: {}
  /api/history:
    get:
      tags:
        - Alerts
      summary: Return alert history of all regions
      parameters:
        - name: from
          in: query
          required: false
          description: Filter from date
          schema:
            type: string
            pattern: ([\d]{4}-[\d]{2}-[\d]{2}T[\d]{2}:[\d]{2}:[\d]{2})
            example: "2020-01-15T05:30:55"
        - name: to
          in: query
          required: false
          description: Filter to date. Default value current date
          schema:
            type: string
            pattern: ([\d]{4}-[\d]{2}-[\d]{2}T[\d]{2}:[\d]{2}:[\d]{2})
            example: "2024-01-15T05:30:55"
        - name: limit
          in: query
          required: false
          description: Limit of result rows. Default value is 1000 rows
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 10

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Status'

  /api/history/{region_id}:
    get:
      tags:
        - Alerts
      summary: Return region alert history
      parameters:
        - name: region_id
          in: path
          required: true
          description: Id of region
          schema:
            type : integer
            format: int64
            minimum: 0
        - name: from
          in: query
          required: false
          description: Filter from date
          schema:
            type: string
            pattern: ([\d]{4}-[\d]{2}-[\d]{2}T[\d]{2}:[\d]{2}:[\d]{2})
            example: "2020-01-15T05:30:55"
        - name: to
          in: query
          required: false
          description: Filter to date. Default value current date
          schema:
            type: string
            pattern: ([\d]{4}-[\d]{2}-[\d]{2}T[\d]{2}:[\d]{2}:[\d]{2})
            example: "2024-01-15T05:30:55"
        - name: limit
          in: query
          required: false
          description: Limit of result rows. Default value is 1000 rows
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 10

      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Status'
        '404':
          description: Error region with given id not found
          content:
            application/json: {}

  /api/renderHtml:
    get:
      tags:
        - Render
      summary: Render Html alert map
      responses:
        '200':
          description: Success
          content:
            text/html: {}
  /api/renderImage:
    get:
      tags:
        - Render
      summary: Render Image alert map
      responses:
        '200':
          description: Success
          content:
            image/png: {}
        '404':
          description: Error image rendering not supported
          content:
            application/json: {}
    
components:
  schemas:
    Region:
      type: object
      properties:
        id:
          type: integer
          example: 10
        name:
          type: string
          example: "Київська"
        is_city:
          type: boolean
          example: true
          
    Status:
      type: object
      properties:
        id:
          type: integer
          example: 5
        region_id:
          type: integer
          example: 10
        is_alert:
          type: boolean
          example: true
        timestamp:
          type: string
          example: "2023-03-12T13:00:28.738938"
          
    ShortStatus:
      type: object
      properties:
        region_id:
          type: integer
          example: 10
        is_alert:
          type: boolean
          example: true
        
  
  